src/requirements.txt
yandex-cloud-ml-sdk==0.15.0
python-telegram-bot==13.15
yandex-speechkit==1.5.0
loguru==0.7.3
boto3==1.40.17
openai==1.99.9